+++ tracy-ui/src/scene/ch06.rs
@@ -43,7 +43,7 @@ impl Scene for PhongSphere {
         "Rendering of a sphere using Phong shading.".to_string()
     }
 
-    fn render(&self, width: usize, height: usize) -> Canvas {
+    fn render(&self, width: u32, height: u32) -> Canvas {
         let mut canvas = Canvas::new(width, height);
         let mut world = World::new();
